+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Add manual align section function
+- Fix enable/disable button on status bar
 
 # v0.0.19 - 2020-01-25
 - Support quoted seperators